haptic/joystick [android]: pass integer (uint16) intensity-values to the java code

+++ src/core/android/SDL_android.h
@@ -88,12 +88,12 @@ int Android_JNI_GetPowerInfo(int *plugged, int *charged, int *battery, int *seco
 /* Joystick support */
 void Android_JNI_JoystickSubscribe(void);
 void Android_JNI_JoystickUnsubscribe(void);
-void Android_JNI_JoystickRumble(int device_id, float low_frequency_intensity, float high_frequency_intensity, int length);
+void Android_JNI_JoystickRumble(int device_id, Uint16 low_frequency_intensity, Uint16 high_frequency_intensity, int length);
 
 /* Haptic support */
 void Android_JNI_HapticSubscribe(void);
 void Android_JNI_HapticUnsubscribe(void);
-void Android_JNI_HapticRun(int device_id, float low_frequency_intensity, float high_frequency_intensity, int length);
+void Android_JNI_HapticRun(int device_id, Uint16 low_frequency_intensity, Uint16 high_frequency_intensity, int length);
 void Android_JNI_HapticStop(int device_id);
 
 /* Video */
